refactor: 중복된 getCategory 로직 공통 폴더인 utils로 분리
