+++ client/src/components/MainCard/MainCard.tsx
@@ -9,6 +9,7 @@ import {
     logoAndTitleContainer,
 } from './MainCard.style';
 import type { MainCardProps } from './types';
+import { getCategory } from '@utils/changeCategory';
 import { Link } from 'react-router-dom';
 
 // FIXME: 나중에 useMemo로 최적화 해야할듯?
@@ -25,24 +26,6 @@ const getTagStatus = (status: string) => {
     }
 };
 
-const getCategory = (category: string) => {
-    switch (category) {
-        case 'PERFORMANCE_ARTS':
-            return '공연동아리';
-        case 'CULTURE':
-            return '문화동아리';
-        case 'SPORTS':
-            return '체육동아리';
-        case 'ACADEMIC':
-            return '학술동아리';
-        case 'VOLUNTEER':
-            return '봉사동아리';
-        case 'RELIGION':
-            return '종교동아리';
-        default:
-            return '미정';
-    }
-};
 function MainCard({
     title = 'En#',
     category = '학술동아리',
